pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "ror-server-bot"
description = "Rigs of Rods Multiplayer Bot for RoRNet 2.44 and Discord"
readme = "README.md"
authors = [{ name = "Danny Mackey", email = "danmack768@gmail.com" }]
requires-python = ">=3.12"
dependencies = [
    "defusedxml~=0.7.1",
    "devtools~=0.12.1",
    "discord.py~=2.3.2",
    "pathvalidate~=3.2.1",
    "pydantic-extra-types~=2.1.0",
    "pydantic~=2.3.0",
    "pyee~=11.0.0",
    "pyyaml~=6.0.2",
    "requests~=2.28.1",
    "rich~=13.5.2",
]
dynamic = ["version"]

[project.urls]
"Source code" = "https://github.com/danmackey/RoRServerBot"

[tool.hatch.envs.default]
dependencies = ["isort", "mypy", "ruff", "tox", "types-pyyaml"]
path = ".venv"

[tool.hatch.publish.index]
disable = true

[tool.hatch.build]
skip-excluded-dirs = true

[tool.hatch.metadata]
allow-direct-references = true

[tool.hatch.version]
path = "ror_server_bot/__init__.py"

[tool.isort]
order_by_type = false
multi_line_output = 3
include_trailing_comma = true

[tool.mypy]
python_version = "3.12"
warn_return_any = true
warn_unused_configs = true
explicit_package_bases = true
disallow_untyped_defs = true
plugins = ["pydantic.mypy"]

[tool.ruff]
line-length = 79
target-version = "py312"

[tool.ruff.lint]
select = [
    "B",    # flake8-bugbear
    "C4",   # flake8-comprehensions
    "C90",  # mccabe
    "E",    # pycodestyle
    "ERA",  # eradicate
    "F",    # Pyflakes
    "ICN",  # flake8-import-conventions
    "ISC",  # flake8-implicit-str-concat
    "N",    # pep8-naming
    "PD",   # pandas-vet
    "PGH",  # pygrep-hooks
    "PIE",  # flake8-pie
    "PLC",  # Pylint
    "PLE",  # Pylint
    "PLR",  # Pylint
    "PLW",  # Pylint
    "PT",   # flake8-pytest-style
    "Q002", # flake8-quotes: BadQuotesDocstring
    "Q003", # flake8-quotes: AvoidQuoteEscape
    "RUF",  # Ruff-specific rules
    "S",    # flake8-bandit
    "SIM",  # flake8-simplify
    "T10",  # flake8-debugger
    "T203", # flake8-print: PPrintFound
    "TID",  # flake8-tidy-imports
    "UP",   # pyupgrade
    "W",    # pycodestyle
    "YTT",  # flake8-2020
]
ignore = [
    "PLR2004", # magic-value-comparison
    "PLR0913", # Too many arguments
    "S101",    # Use of assert detected
    "SIM108",  # Use ternary operator {contents} instead of if-else-block
    "SIM110",  # Use {repl} instead of for loop
    "SIM117",  # Use a single with statement with multiple contexts instead of nested with statements
    "UP015",   # Unnecessary open mode parameters
]


[tool.ruff.lint.pep8-naming]
classmethod-decorators = [
    "classmethod",
    "pydantic.validator",
    "pydantic.field_validator"
]
